update - move custom thread logic into own controller
